+++ Source/API/Refund.swift
@@ -84,7 +84,7 @@ public class Refund: NSObject {
     
     - Returns: reactive self
     */
-    public func completion(block: (Response?, JudoError?) -> ()) -> Self {
+    public func completion(block: JudoCompletionBlock) -> Self {
         
         let parameters = Session.progressionParameters(self.receiptID, amount: self.amount, paymentReference: self.paymentReference)