+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Released on 2016-01-14
 - Card number information to CardDetails object
 - Package file for future Swift package manager
 - 100% Documented
+- Updated Documentation and added a lot of new snippets
+- Response class now conforms to SequenceType, CollectionType, GeneratorType and ArrayLiteralConvertible
 
 #### Changed